chore/#139: 더미데이터 파일 분리
